Move env variable calling to viper
