+++ .envrc
@@ -4,6 +4,10 @@
 # Project Configuration #
 #########################
 
+export CLIENT_HOSTNAME=http://localhost
+export CLIENT_PORT=3000
+export CLIENT_ADDRESS=$CLIENT_HOSTNAME:$CLIENT_PORT
+
 # OKTA variables
 
 export OKTA_CLIENT_ID=0oa2e913coDQeG19S297
