+++ cmd/easi/serve.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 
 	"github.com/cmsgov/easi-app/pkg/server"
 )
@@ -11,6 +12,7 @@ var serveCmd = &cobra.Command{
 	Short: "Serve the EASi application",
 	Long:  `Serve the EASi application`,
 	Run: func(cmd *cobra.Command, args []string) {
+		viper.AutomaticEnv()
 		server.Serve()
 	},
 }
