+++ go.mod
@@ -18,6 +18,6 @@ require (
 	github.com/spf13/cobra v0.0.5
 	github.com/spf13/jwalterweatherman v1.1.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
-	github.com/spf13/viper v1.4.0 // indirect
+	github.com/spf13/viper v1.4.0
 	golang.org/x/sys v0.0.0-20191028164358-195ce5e7f934 // indirect
 )
