+++ pkg/server/server_test.go
@@ -1,11 +1,1 @@
 package server
-
-import "testing"
-
-func TestVerify(t *testing.T) {
-	verifybool := isAuthenticated(
-		"Bearer eyJraWQiOiJEUlpHd1VDRVNCNGRGLTR1WW85UE9sNy1nb3drMlhnSF9oem5JUzc0TndNIiwiYWxnIjoiUlMyNTYifQ.eyJ2ZXIiOjEsImp0aSI6IkFULndaTExMQ2RkX2NHTW9oRmVzSEVKZFE4ZHZYYlNQOW9Fajlvb1l6a1NwTkkiLCJpc3MiOiJodHRwczovL3Rlc3QuaWRwLmlkbS5jbXMuZ292IiwiYXVkIjoiaHR0cHM6Ly90ZXN0LmlkcC5pZG0uY21zLmdvdiIsInN1YiI6IkxGSUQiLCJpYXQiOjE1NzYwMjY3MTIsImV4cCI6MTU3NjAzMDMxMiwiY2lkIjoiMG9hMmU5MTNjb0RRZUcxOVMyOTciLCJ1aWQiOiIwMHUya3ozN21iMkpuSENQZTI5NyIsInNjcCI6WyJvcGVuaWQiLCJlbWFpbCJdfQ.kEnw_tT8bXwQICkztSkPaId2zh3m7N31PfReQzA2Q7QcSzjXFamaRNJboLK7LwV0Y0YyPOQloahFR7oVsxkJj0uF9aom1-Q9D3yEQAyPLuFy_BNb7b0kzeyIy9AlkFjFmJINTbjXiF06s3kfXNzrX_h3YZPomtZE61QWMIco1Z8pflvYbUaxePM7oTZ6vZgK6U8WaeKzraXLOUEMxDox8FFlcn7zcA6In1IJRe5j6q0m8aim3kfntYGlbrc_gRWQ5H9pHTmrR4jbj5WMnJ62WTIegieXUeKWdBwhH-wpX-KAzFvdAwg2qh3AGuQbKz5t_RzhBFWnNbdLh88NBTwf-Q")
-	if verifybool != true {
-		t.Errorf("I didn't isAuthenticated!")
-	}
-}
